requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0

# CLIP
open-clip-torch>=2.20.0

# Grounding DINO
groundingdino-py>=0.4.0

# Vector Database
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Use this if you have GPU

# Database
mysql-connector-python>=8.0.33
pymysql>=1.1.0
sqlalchemy>=2.0.0

# API & HTTP
requests>=2.31.0
httpx>=0.24.1
aiohttp>=3.8.0
aiomysql>=0.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart>=0.0.6

# AWS S3 (for vector database download)
boto3>=1.28.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
pandas>=2.0.0
Pillow>=10.0.0

# Image Processing
albumentations>=1.3.1

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
ipython>=8.14.0
jupyter>=1.0.0
